Release 0.1.20140613 - disable armv64 (will enable on next release) - fix release warnings (common lib-related will be fixed on next release)

+++ iOS/SmartDeviceApp/Classes/Helpers/PDFRenderOperation.m
@@ -531,7 +531,7 @@
     CGContextSaveGState(contextRef);
     
     CGContextSetLineWidth(contextRef, 2.0f);
-    float dashLine[] = { 16.0f, 16.0f };
+    CGFloat dashLine[] = { 16.0f, 16.0f };
     
     // Draw BG line (white)
     CGContextSetLineDash(contextRef, 16.0f, dashLine, 2);
